backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Patryk Wójtowicz", email = "patryk.zbigniew.wojtowicz@gmail.com" }
]
requires-python = ">=3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12.0"
fastapi = "^0.115.3"
uvicorn = "^0.34.0"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"
psycopg = {extras = ["binary"], version = "^3.2.0"}
python-dotenv = "^1.0.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
urllib3 = "^2.4.0"
python-dateutil = "^2.9.0.post0"
typing-extensions = "^4.13.2"
tenacity = "^9.1.2"
httpx = "^0.28.1"
certifi = "^2026.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.12.1"
pylint = "^3.3.1"
bandit = "^1.7.10"
black = "^24.10.0"
isort = "^5.13.2"
pytest-postgresql = "^8.0.0"
pytest-env = "^1.1.0"
factory-boy = "^3.3.3"
pytest-factoryboy = "^2.8.1"
pytest-cov = "^7.0.0"

[tool.poetry.scripts]
"db:seed" = "scripts.db:seed"
"db:clear" = "scripts.db:clear"
lint = "scripts.utils:lint"
format = "scripts.utils:format_"
"wl:generate-client" = "scripts.whitelabel:generate_client"
"wl:retry-sync" = "scripts.whitelabel:retry_sync"


[tool.pytest.ini_options]
env = [
    "API_KEY=test-api-key",
    "DB_NAME=postgres_test",
]
addopts = "--cov=backend --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "backend/migrations/*",
    "backend/whitelabel_client/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = "missing-module-docstring, missing-function-docstring, missing-class-docstring, fixme"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.bandit]
skips = ["B101"]  # Example: Skip assert statements warnings

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
